requirements.txt
scapy
pandas
plotly
pyshark
python-dotenv
openai
groq
flask
flask-cors
